==> 1 - Gold.ipynb <==
import pandas as pd
import numpy as np

from sklearn.preprocessing import StandardScaler
from sklearn.grid_search import GridSearchCV
from sklearn.cross_validation import KFold

from sklearn.cross_validation import cross_val_score

from sklearn.linear_model import LogisticRegression
from sklearn.ensemble import RandomForestClassifier
from sklearn.linear_model import Lasso, Ridge

%pylab inline

def normalize_data(X):
    return pd.DataFrame(
        StandardScaler().fit_transform(X),
        index = X.index, 
        columns=X.columns
    )

def logEstimation(X, y):
    grid = {'C': np.power(10.0, np.arange(-5, 1))}
    kf=KFold(y.size, n_folds=5, shuffle=True, random_state=241)
    clf=LogisticRegression(random_state=241)
    gs = GridSearchCV(clf, grid, scoring='roc_auc', cv=kf)
    gs.fit(X, y)
    return gs

train_matches = pd.read_csv('data/train.csv')
test_matches = pd.read_csv('data/test.csv')

gold = pd.read_csv('data/gold.csv', index_col='mid')
gold = gold[gold.times == 600]
gold.drop('times', 1, inplace=True)

gold.head()

# Preprocessing

norm_gold = pd.DataFrame(index=gold.index)
for col in gold.columns:
    norm_gold['gold_' + col] = gold[col]
norm_gold.head()

radiant_gold = gold[['player_0', 'player_1', 'player_2', 'player_3', 'player_4']].sum(axis=1)
dire_gold = gold[['player_5', 'player_6', 'player_7', 'player_8', 'player_9']].sum(axis=1)
radiant_carry_gold = gold[['player_0', 'player_1', 'player_2', 'player_3', 'player_4']].max(axis=1)
dire_carry_gold = gold[['player_5', 'player_6', 'player_7', 'player_8', 'player_9']].max(axis=1)

norm_gold['radiant_gold'] = radiant_gold
norm_gold['dire_gold'] = dire_gold
norm_gold['radiant_carry_gold'] = radiant_carry_gold
norm_gold['dire_carry_gold'] = dire_carry_gold
# norm_gold['diff_gold'] = radiant_gold - dire_gold
norm_gold.head()

norm_gold = normalize_data(norm_gold)
norm_gold.head()

norm_gold['ratio_gold'] = radiant_gold / dire_gold
norm_gold.head()

norm_gold = norm_gold.reset_index()
norm_gold.head()

norm_gold.to_csv('processing_tables/norm_gold.csv', index=None)

# New features

new_gold = pd.read_csv('data/gold.csv', index_col='mid')
new_gold = new_gold[new_gold.times == 600]
new_gold.drop('times', 1, inplace=True)
new_gold.head()

heroes = pd.read_csv('data/heroes.csv', index_col='mid')
heroes.head()

heroes.max()[0]

mean_gold = np.zeros(111)
for hero_num in range(111):
    hero_values = []
    for player_num in range(10):
        colname = 'player_{}'.format(player_num)
        hero_index = heroes.loc[heroes[colname] == hero_num].index
        hero_values.extend(new_gold.loc[hero_index][colname].values)
    mean_gold[hero_num] = int(np.mean(hero_values))

mean_gold[0] = 100000
print mean_gold

mean_gold == 0 - соответствует вышедшему из игры человеку

mean_predict_gold = heroes.apply(lambda x: mean_gold[x])
mean_predict_gold.head()

temp_gold_score = new_gold / mean_predict_gold
gold_score_by_mean = pd.DataFrame(index=temp_gold_score.index)
for col in temp_gold_score.columns:
#     gold_score_by_mean['gold_score_' + col] = StandardScaler().fit_transform(temp_gold_score[col].reshape(-1, 1))
    gold_score_by_mean['gold_score_' + col] = temp_gold_score[col]

# gold_score_by_mean['radiant_norm_gold'] = norm_gold.radiant_gold
# gold_score_by_mean['dire_norm_gold'] = norm_gold.dire_gold
# gold_score_by_mean['radiant_carry_norm_gold'] = norm_gold.radiant_carry_gold
# gold_score_by_mean['dire_carry_norm_gold'] = norm_gold.dire_carry_gold

gold_score_by_mean['radiant_best_gold_score'] = gold_score_by_mean.iloc[:,1:6].max(1)
gold_score_by_mean['dire_best_gold_score'] = gold_score_by_mean.iloc[:,6:11].max(1)
gold_score_by_mean['radiant_sum_gold_score'] = StandardScaler().fit_transform(gold_score_by_mean.iloc[:,1:6].sum(1).reshape(-1, 1))
gold_score_by_mean['dire_sum_gold_score'] = StandardScaler().fit_transform(gold_score_by_mean.iloc[:,6:11].sum(1).reshape(-1, 1))

# gold_score_by_mean = normalize_data(gold_score_by_mean)
gold_score_by_mean.reset_index(level=0, inplace=True)
gold_score_by_mean.head()

norm_gold.head()

pd.merge(gold_score_by_mean, norm_gold, on='mid').to_csv('processing_tables/old_gold.csv', index=None)

gold_score_by_mean.to_csv('processing_tables/goldScoreWithCarry.csv', index=None)

cropped_gold_score = gold_score_by_mean.copy()
cropped_gold_score.drop(['radiant_carry_norm_gold', 'dire_carry_norm_gold'], 1)
cropped_gold_score.head()
cropped_gold_score.to_csv('processing_tables/goldScoreWithoutCarry.csv', index=None)

# Test only norm gold

train = pd.merge(train_matches, norm_gold, on='mid', how='left')
test = pd.merge(test_matches, norm_gold, on='mid', how='left')
X_train = train.drop('radiant_won', 1)
y_train = train.radiant_won

norm_gold.to_csv('processing_tables/norm_gold.csv', index=None)
train.to_csv('processing_tables/train_carry_gold.csv', index=None)
test.to_csv('processing_tables/test_carry_gold.csv', index=None)

clf = RandomForestClassifier(n_estimators=100, n_jobs=-1, random_state=228)
np.mean(cross_val_score(clf, X_train, y_train, cv=5, scoring='roc_auc'))

gs = logEstimation(X_train, y_train)
print gs.grid_scores_
print gs.best_params_
print('best score: {}'.format(gs.best_score_))

clf=LogisticRegression(random_state=241, C=0.01)
np.mean(cross_val_score(clf, X_train, y_train, cv=5, scoring='roc_auc'))

clf=Ridge(alpha=0.0001)
np.mean(cross_val_score(clf, X_train, y_train, cv=5, scoring='roc_auc'))

clf=Lasso(alpha=0.0001, max_iter=6000)
np.mean(cross_val_score(clf, X_train, y_train, cv=5, scoring='roc_auc'))

# Test only gold score

train = pd.merge(train_matches, gold_score_by_mean, on='mid', how='left')
test = pd.merge(test_matches, gold_score_by_mean, on='mid', how='left')
X_train = train.drop('radiant_won', 1)
y_train = train.radiant_won

clf = RandomForestClassifier(n_estimators=100, n_jobs=-1, random_state=228)
np.mean(cross_val_score(clf, X_train, y_train, cv=5, scoring='roc_auc'))

gs = logEstimation(X_train, y_train)
print gs.grid_scores_
print gs.best_params_
print('best score: {}'.format(gs.best_score_))

clf=LogisticRegression(random_state=241, C=0.1)
np.mean(cross_val_score(clf, X_train, y_train, cv=5, scoring='roc_auc'))

clf=Ridge(alpha=0.0001)
np.mean(cross_val_score(clf, X_train, y_train, cv=5, scoring='roc_auc'))

clf=Lasso(alpha=0.0001, max_iter=4000)
np.mean(cross_val_score(clf, X_train, y_train, cv=5, scoring='roc_auc'))

train.to_csv('processing_tables/train_score_gold.csv', index=None)
test.to_csv('processing_tables/test_score_gold.csv', index=None)
gold_score_by_mean.to_csv('processing_tables/gold_scores.csv', index=None)

# RANKS

## NormByRowsRanks

gold = pd.read_csv('data/gold.csv', index_col='mid')
gold = gold[gold.times == 600]
gold.drop('times', 1, inplace=True)

radiant_rank_gold = pd.DataFrame(data = np.sort(gold.values[:,:5]),
                                 index=gold.index,
                                 columns = ['radiant_gold_rank_{0}'.format(i) for i in range(5)]) 
# radiant_rank_gold = normalize_data(radiant_rank_gold)

radiant_rank_gold.reset_index(inplace=True)
radiant_rank_gold.head()

dire_rank_gold = pd.DataFrame(data = np.sort(gold.values[:,5:10]),
                                 index=gold.index,
                                 columns = ['dire_gold_rank_{0}'.format(i) for i in range(5)]) 
# dire_rank_gold = normalize_data(dire_rank_gold)

dire_rank_gold.reset_index(inplace=True)

dire_rank_gold.head()

rank_gold = pd.merge(radiant_rank_gold, dire_rank_gold, on='mid', how='left')
rank_gold.head()

norm_rank_gold = pd.DataFrame(rank_gold.mid)
row_sum = rank_gold.drop('mid', 1).sum(1)
for col in rank_gold.columns[1:]:
    norm_rank_gold[col] = rank_gold[col] / row_sum
norm_rank_gold.ix[:,1:] = normalize_data(norm_rank_gold.drop('mid', 1))
norm_rank_gold.head()

norm_rank_gold.to_csv('processing_tables/normByRowsRankGold.csv', index=None)

train = pd.merge(norm_rank_gold, train_matches, on='mid', how='right')
test = pd.merge(norm_rank_gold, test_matches, on='mid', how='right')
X_train = train.drop('radiant_won', 1)
y_train = train.radiant_won

clf = RandomForestClassifier(n_estimators=100, n_jobs=-1, random_state=228)
np.mean(cross_val_score(clf, X_train, y_train, cv=5, scoring='roc_auc'))

gs = logEstimation(X_train, y_train)
print gs.grid_scores_
print gs.best_params_
print('best score: {}'.format(gs.best_score_))

clf=LogisticRegression(random_state=241, C=0.001)
np.mean(cross_val_score(clf, X_train, y_train, cv=5, scoring='roc_auc'))

clf=Ridge(alpha=0.0001)
np.mean(cross_val_score(clf, X_train, y_train, cv=5, scoring='roc_auc'))

# NormByRowsRanks and scores

train = pd.merge(norm_rank_gold, train_matches, on='mid', how='right')
test = pd.merge(norm_rank_gold, test_matches, on='mid', how='right')

train = pd.merge(train, gold_score_by_mean, on='mid', how='left')
test = pd.merge(test, gold_score_by_mean, on='mid', how='left')

X_train = train.drop(['radiant_carry_norm_gold', 'dire_carry_norm_gold'], 1)
X_test = test.drop(['radiant_carry_norm_gold', 'dire_carry_norm_gold'], 1)

X_train = train.drop('radiant_won', 1)
y_train = train.radiant_won

clf = RandomForestClassifier(n_estimators=100, n_jobs=-1, random_state=228)
np.mean(cross_val_score(clf, X_train, y_train, cv=5, scoring='roc_auc'))

gs = logEstimation(X_train, y_train)
print gs.grid_scores_
print gs.best_params_
print('best score: {}'.format(gs.best_score_))

clf=LogisticRegression(random_state=241, C=0.01)
np.mean(cross_val_score(clf, X_train, y_train, cv=5, scoring='roc_auc'))

clf=Ridge(alpha=0.0001)
np.mean(cross_val_score(clf, X_train, y_train, cv=5, scoring='roc_auc'))

clf=Lasso(alpha=0.0001, max_iter=6000)
np.mean(cross_val_score(clf, X_train, y_train, cv=5, scoring='roc_auc'))

## Just rank gold and gold scores

rank_gold.head()

stupid_norm_rank_gold = rank_gold.copy()
stupid_norm_rank_gold.ix[:,1:] = normalize_data(rank_gold.drop('mid', 1))
stupid_norm_rank_gold.head()

stupid_norm_rank_gold.to_csv('processing_tables/normByColsRankGold.csv', index=None)

train = pd.merge(train_matches, stupid_norm_rank_gold, on='mid', how='left')
test = pd.merge(test_matches, stupid_norm_rank_gold, on='mid', how='left')
train = pd.merge(train, gold_score_by_mean, on='mid', how='left')
test = pd.merge(test, gold_score_by_mean, on='mid', how='left')
X_train = train.drop(['radiant_carry_norm_gold', 'dire_carry_norm_gold'], 1)
X_test = test.drop(['radiant_carry_norm_gold', 'dire_carry_norm_gold'], 1)

X_train = train.drop('radiant_won', 1)
y_train = train.radiant_won

clf = RandomForestClassifier(n_estimators=100, n_jobs=-1, random_state=228)
np.mean(cross_val_score(clf, X_train, y_train, cv=5, scoring='roc_auc'))

gs = logEstimation(X_train, y_train)
print gs.grid_scores_
print gs.best_params_
print('best score: {}'.format(gs.best_score_))

clf=LogisticRegression(random_state=241, C=0.1)
np.mean(cross_val_score(clf, X_train, y_train, cv=5, scoring='roc_auc'))

clf=Ridge(alpha=0.0001)
np.mean(cross_val_score(clf, X_train, y_train, cv=5, scoring='roc_auc'))

clf=Lasso(alpha=0.0001, max_iter=4000)
np.mean(cross_val_score(clf, X_train, y_train, cv=5, scoring='roc_auc'))

train.to_csv('processing_tables/train_rank_score_gold.csv', index=None)
test.to_csv('processing_tables/test_rank_score_gold.csv', index=None)
gold_score_by_mean.to_csv('processing_tables/stupidRank+gold_scores.csv', index=None)

# Diff rank

gold_rank_dif = pd.DataFrame(data = np.sort(gold.values[:,:5]) - np.sort(gold.values[:,5:10]),
                                 index=gold.index,
                                 columns = ['gold_diff_rank_{0}'.format(i) for i in range(5)]) 
gold_rank_dif = normalize_data(gold_rank_dif)

gold_rank_dif.reset_index(inplace=True)
gold_rank_dif.head()

gold_rank_dif.to_csv('processing_tables/goldRankDif.csv', index=None)

train = pd.merge(train_matches, gold_rank_dif, on='mid', how='left')
test = pd.merge(test_matches, gold_rank_dif, on='mid', how='left')
train = pd.merge(train, gold_score_by_mean, on='mid', how='left')
test = pd.merge(test, gold_score_by_mean, on='mid', how='left')

X_train = train.drop(['radiant_carry_norm_gold', 'dire_carry_norm_gold'], 1)
X_test = test.drop(['radiant_carry_norm_gold', 'dire_carry_norm_gold'], 1)

X_train = train.drop(['radiant_won'], 1)
y_train = train.radiant_won

X_train.head()

train.to_csv('processing_tables/train_scoreDiffRankGold.csv', index=None)
test.to_csv('processing_tables/test_scoreDiffRankGold.csv', index=None)

clf = RandomForestClassifier(n_estimators=100, n_jobs=-1, random_state=228)
np.mean(cross_val_score(clf, X_train, y_train, cv=5, scoring='roc_auc'))

gs = logEstimation(X_train, y_train)
print gs.grid_scores_
print gs.best_params_
print('best score: {}'.format(gs.best_score_))

clf=LogisticRegression(random_state=241, C=0.01)
np.mean(cross_val_score(clf, X_train, y_train, cv=5, scoring='roc_auc'))

clf=Ridge(alpha=0.0001)
np.mean(cross_val_score(clf, X_train, y_train, cv=5, scoring='roc_auc'))

clf=Lasso(alpha=0.0001, max_iter=4000)
np.mean(cross_val_score(clf, X_train, y_train, cv=5, scoring='roc_auc'))

## Diff rank rates

rank_gold.head()

norm_by_row_gold_rank_dif = pd.DataFrame(data = np.sort(gold.values[:,:5]) - np.sort(gold.values[:,5:10]),
                                 index=gold.index,
                                 columns = ['gold_diff_rank_{0}'.format(i) for i in range(5)]) 
norm_by_row_gold_rank_dif.head()

norm_by_row_dif_gold =
