==> 1 - gold_from_scratch.ipynb <==
import pandas as pd
import numpy as np

from sklearn.preprocessing import StandardScaler
from sklearn.grid_search import GridSearchCV
from sklearn.cross_validation import KFold

from sklearn.cross_validation import cross_val_score

from sklearn.linear_model import LogisticRegression
from sklearn.ensemble import RandomForestClassifier
from sklearn.linear_model import Lasso, Ridge

%pylab inline

def normalize_data(X):
    return pd.DataFrame(
        StandardScaler().fit_transform(X),
        index = X.index, 
        columns=X.columns
    )

def logEstimation(X, y):
    grid = {'C': np.power(10.0, np.arange(-5, 6))}
    kf=KFold(y.size, n_folds=5, shuffle=True, random_state=241)
    clf=LogisticRegression(random_state=241)
    gs = GridSearchCV(clf, grid, scoring='roc_auc', cv=kf)
    gs.fit(X, y)
    return gs

train_matches = pd.read_csv('data/train.csv')
test_matches = pd.read_csv('data/test.csv')

gold = pd.read_csv('data/gold.csv', index_col='mid')
gold = gold[gold.times == 600]
gold.drop('times', 1, inplace=True)

norm_gold = pd.DataFrame(index=gold.index)
for col in gold.columns:
    norm_gold['gold_' + col] = gold[col]
norm_gold.head()

radiant_gold = gold[['player_0', 'player_1', 'player_2', 'player_3', 'player_4']].sum(axis=1)
dire_gold = gold[['player_5', 'player_6', 'player_7', 'player_8', 'player_9']].sum(axis=1)
radiant_carry_gold = gold[['player_0', 'player_1', 'player_2', 'player_3', 'player_4']].max(axis=1)
dire_carry_gold = gold[['player_5', 'player_6', 'player_7', 'player_8', 'player_9']].max(axis=1)
norm_gold = normalize_data(norm_gold)
norm_gold.reset_index(inplace=True)
norm_gold.head()

train = pd.merge(train_matches, norm_gold, on='mid', how='left')
test = pd.merge(test_matches, norm_gold, on='mid', how='left')
X_train = train.drop(['radiant_won', 'mid'], 1)
y_train = train.radiant_won

# Usual_gold_score

new_gold = pd.read_csv('data/gold.csv', index_col='mid')
new_gold = new_gold[new_gold.times == 600]
new_gold.drop('times', 1, inplace=True)
heroes = pd.read_csv('data/heroes.csv', index_col='mid')

mean_gold = np.zeros(111)
for hero_num in range(111):
    hero_values = []
    for player_num in range(10):
        colname = 'player_{}'.format(player_num)
        hero_index = heroes.loc[heroes[colname] == hero_num].index
        hero_values.extend(new_gold.loc[hero_index][colname].values)
    mean_gold[hero_num] = int(np.mean(hero_values))

mean_gold = new_mean_gold
mean_gold[0] = 100000

mean_predict_gold = heroes.apply(lambda x: mean_gold[x])

temp_gold_score = new_gold / mean_predict_gold
usual_gold_score = pd.DataFrame(index=temp_gold_score.index)
for col in temp_gold_score.columns:
    usual_gold_score['usual_gold_score_' + col] = temp_gold_score[col]
usual_gold_score.reset_index(level=0, inplace=True)

radiant_rank_gold_score = pd.DataFrame(data = np.sort(usual_gold_score.values[:,1:6]),
                                       index=adapt_gold_score.index,
                                       columns = ['gold_score_rank_{0}'.format(i) for i in range(5)]) 

dire_rank_gold_score = pd.DataFrame(data = np.sort(usual_gold_score.values[:,6:11]),
                                       index=adapt_gold_score.index,
                                       columns = ['gold_score_rank_{0}'.format(i) for i in range(5)]) 

gold_rank_dif = pd.DataFrame(data = np.sort(gold.values[:,0:5]) - np.sort(gold.values[:,5:10]),
                             index=gold.index,
                             columns = ['gold_rank_{}'.format(i) for i in range(5)]) 

gold_stats_usual = pd.DataFrame(index=gold.index)
gold_stats_usual['gold_dif'] = radiant_gold - dire_gold
gold_stats_usual['top_gold_dif'] = gold_rank_dif.gold_rank_4
gold_stats_usual['second_gold_dif'] = gold_rank_dif.gold_rank_3
# gold_stats_usual['third_gold_dif'] = gold_rank_dif.gold_rank_2
gold_stats_usual['sum_usual_gold_score_dif'] = radiant_rank_gold_score.sum(1) - dire_rank_gold_score.sum(1)
gold_stats_usual['top_usual_gold_score_dif'] = radiant_rank_gold_score.gold_score_rank_4 - dire_rank_gold_score.gold_score_rank_4
gold_stats_usual['second_usual_gold_score_dif'] = radiant_rank_gold_score.gold_score_rank_3 - dire_rank_gold_score.gold_score_rank_3
gold_stats_usual['third_usual_gold_score_dif'] = radiant_rank_gold_score.gold_score_rank_2 - dire_rank_gold_score.gold_score_rank_2
gold_stats_usual = normalize_data(gold_stats_usual)
gold_stats_usual.reset_index(inplace=True)
gold_stats_usual.to_csv('processing_tables/gold_stats_usual_dif.csv', index=None)
gold_stats_usual.head()



# adaptive gold_score

new_gold = pd.read_csv('data/gold.csv', index_col='mid')
new_gold = new_gold[new_gold.times == 600]
new_gold.drop('times', 1, inplace=True)
heroes = pd.read_csv('data/heroes.csv', index_col='mid')

mean_gold = np.zeros(111)
for hero_num in range(111):
    hero_values = []
    for player_num in range(10):
        colname = 'player_{}'.format(player_num)
        hero_index = heroes.loc[heroes[colname] == hero_num].index
        hero_values.extend(new_gold.loc[hero_index][colname].values)
    mean_gold[hero_num] = int(np.mean(hero_values))

print 'Source mean gold:\n', mean_gold
print 'Sorted mean gold:\n', np.sort(mean_gold)
new_mean_gold = np.array(map(lambda x: x if x > 4500 else 4500 + 5*(4500 - x), mean_gold))
# new_mean_gold = np.array(map(lambda x: x if x > 4500 else 4500 + int((4500 - x)*2 / 1000), mean_gold))
mean_gold = new_mean_gold
mean_gold[0] = 100000
print 'After threshold:\n', mean_gold

mean_predict_gold = heroes.apply(lambda x: mean_gold[x])
# mean_predict_gold.head()

temp_gold_score = new_gold / mean_predict_gold
adapt_gold_score = pd.DataFrame(index=temp_gold_score.index)
for col in temp_gold_score.columns:
    adapt_gold_score['adapt_gold_score_' + col] = temp_gold_score[col]
adapt_gold_score.reset_index(level=0, inplace=True)

radiant_rank_gold_score = pd.DataFrame(data = np.sort(adapt_gold_score.values[:,1:6]),
                                       index=adapt_gold_score.index,
                                       columns = ['gold_score_rank_{0}'.format(i) for i in range(5)]) 
# radiant_rank_gold_score = normalize_data(radiant_rank_gold_score)
dire_rank_gold_score = pd.DataFrame(data = np.sort(adapt_gold_score.values[:,6:11]),
                                       index=adapt_gold_score.index,
                                       columns = ['gold_score_rank_{0}'.format(i) for i in range(5)]) 
# dire_rank_gold_score = normalize_data(dire_rank_gold_score)
dire_rank_gold_score.head()
# print 

gold_rank_dif = pd.DataFrame(data = np.sort(gold.values[:,0:5]) - np.sort(gold.values[:,5:10]),
                             index=gold.index,
                             columns = ['gold_rank_{}'.format(i) for i in range(5)]) 

gold_stats_adapt = pd.DataFrame(index=gold.index)
gold_stats_adapt['gold_dif'] = radiant_gold - dire_gold
gold_stats_adapt['top_gold_dif'] = gold_rank_dif.gold_rank_4
gold_stats_adapt['second_gold_dif'] = gold_rank_dif.gold_rank_3
# gold_stats_adapt['third_gold_dif'] = gold_rank_dif.gold_rank_2
gold_stats_adapt['adapt_score_sum_dif'] = radiant_rank_gold_score.sum(1) - dire_rank_gold_score.sum(1)
gold_stats_adapt['top_adapt_gold_score_dif'] = radiant_rank_gold_score.gold_score_rank_4 - dire_rank_gold_score.gold_score_rank_4
gold_stats_adapt['second_adapt_gold_score_dif'] = radiant_rank_gold_score.gold_score_rank_3 - dire_rank_gold_score.gold_score_rank_3
gold_stats_adapt['third_adapt_gold_score_dif'] = radiant_rank_gold_score.gold_score_rank_2 - dire_rank_gold_score.gold_score_rank_2
gold_stats_adapt = normalize_data(gold_stats_adapt)
gold_stats_adapt.reset_index(inplace=True)
gold_stats_adapt.to_csv('processing_tables/gold_stats_adapt_dif.csv', index=None)
gold_stats_adapt.head()

# threshold gold score

new_gold = pd.read_csv('data/gold.csv', index_col='mid')
new_gold = new_gold[new_gold.times == 600]
new_gold.drop('times', 1, inplace=True)
heroes = pd.read_csv('data/heroes.csv', index_col='mid')

mean_gold = np.zeros(111)
for hero_num in range(111):
    hero_values = []
    for player_num in range(10):
        colname = 'player_{}'.format(player_num)
        hero_index = heroes.loc[heroes[colname] == hero_num].index
        hero_values.extend(new_gold.loc[hero_index][colname].values)
    mean_gold[hero_num] = int(np.mean(hero_values))

mean_gold[0] = 100000
print 'Source:\n', mean_gold
print 'Sorted:\n', np.sort(mean_gold)
new_mean_gold = np.array(map(lambda x: x if x > 4400 else 10000, mean_gold))
mean_gold = new_mean_gold
print 'After threshold:\n', mean_gold

mean_predict_gold = heroes.apply(lambda x: mean_gold[x])
mean_predict_gold.head()

temp_gold_score = new_gold / mean_predict_gold
adapt_gold_score = pd.DataFrame(index=temp_gold_score.index)
for col in temp_gold_score.columns:
    adapt_gold_score['adapt_gold_score_' + col] = temp_gold_score[col]
adapt_gold_score.reset_index(level=0, inplace=True)

radiant_rank_gold_score = pd.DataFrame(data = np.sort(gold_score_by_mean.values[:,1:6]),
                                       index=gold_score_by_mean.index,
                                       columns = ['gold_score_rank_{0}'.format(i) for i in range(5)]) 
# radiant_rank_gold_score = normalize_data(radiant_rank_gold_score)
dire_rank_gold_score = pd.DataFrame(data = np.sort(gold_score_by_mean.values[:,6:11]),
                                       index=gold_score_by_mean.index,
                                       columns = ['gold_score_rank_{0}'.format(i) for i in range(5)]) 
# dire_rank_gold_score = normalize_data(dire_rank_gold_score)
dire_rank_gold_score.head()

gold_rank_dif = pd.DataFrame(data = np.sort(gold.values[:,0:5]) - np.sort(gold.values[:,5:10]),
                             index=gold.index,
                             columns = ['gold_rank_{}'.format(i) for i in range(5)]) 

gold_stats_threshold = pd.DataFrame(index=new_gold.index)
gold_stats_threshold['gold_dif'] = radiant_gold - dire_gold
gold_stats_threshold['top_gold_dif'] = gold_rank_dif.gold_rank_4
gold_stats_threshold['second_gold_dif'] = gold_rank_dif.gold_rank_3
gold_stats_threshold['third_gold_dif'] = gold_rank_dif.gold_rank_2
gold_stats_threshold['threshold_score_sum_dif'] = radiant_rank_gold_score.sum(1) - dire_rank_gold_score.sum(1)
gold_stats_threshold['top_threshold_gold_score_dif'] = radiant_rank_gold_score.gold_score_rank_4 - dire_rank_gold_score.gold_score_rank_4
gold_stats_threshold['second_threshold_gold_score_dif'] = radiant_rank_gold_score.gold_score_rank_3 - dire_rank_gold_score.gold_score_rank_3
gold_stats_threshold['third_threshold_gold_score_dif'] = radiant_rank_gold_score.gold_score_rank_2 - dire_rank_gold_score.gold_score_rank_2
gold_stats_threshold = normalize_data(gold_stats_threshold)
gold_stats_threshold.reset_index(inplace=True)
gold_stats_threshold.to_csv('processing_tables/gold_stats_threshold_dif.csv', index=None)
gold_stats_threshold.head()

gold_stats = pd.DataFrame(index=gold.index)
gold_stats['radiant_gold'] = radiant_gold
gold_stats['dire_gold'] = dire_gold
gold_stats['radiant_top_gold'] = rank_gold.radiant_gold_rank_4
gold_stats['radiant_second_gold'] = rank_gold.radiant_gold_rank_3
gold_stats['dire_top_gold'] = rank_gold.dire_gold_rank_4
gold_stats['dire_second_gold'] = rank_gold.dire_gold_rank_3
# gold_stats = normalize_data(gold_stats)
gold_stats.reset_index(inplace=True)

gold_stats['radiant_top_gold_score'] = radiant_rank_gold_score.radiant_gold_score_rank_4
gold_stats['radiant_second_gold_score'] = radiant_rank_gold_score.radiant_gold_score_rank_3
gold_stats['dire_top_gold_score'] = dire_rank_gold_score.dire_gold_score_rank_4
gold_stats['dire_second_gold_score'] = dire_rank_gold_score.dire_gold_score_rank_3
gold_stats.head()

train = pd.merge(train_matches, gold_stats, on='mid', how='left')
test = pd.merge(test_matches, gold_stats, on='mid', how='left')
X_train = train.drop(['radiant_won', 'mid'], 1)
y_train = train.radiant_won
X_train = normalize_data(X_train)

clf=LogisticRegression(random_state=241, C=0.1)
print np.mean(cross_val_score(clf, X_train, y_train, cv=3, scoring='roc_auc'))
clf=Ridge(alpha=0.0001)
print np.mean(cross_val_score(clf, X_train, y_train, cv=5, scoring='roc_auc'))
clf=Lasso(alpha=0.0001, max_iter=6000)
print np.mean(cross_val_score(clf, X_train, y_train, cv=5, scoring='roc_auc'))

gold_stats.to_csv('processing_tables/gold_stat.csv', index=None)

gold_stats_dif = pd.DataFrame(index=gold.index)
gold_stats_dif['sum_gold_dif'] = radiant_gold - dire_gold
gold_stats_dif['gold_top_dif'] = rank_gold.radiant_gold_rank_4 - rank_gold.dire_gold_rank_4
gold_stats_dif['gold_second_dif'] = rank_gold.radiant_gold_rank_3 - rank_gold.dire_gold_rank_3
gold_stats_dif = normalize_data(gold_stats_dif)

gold_stats_dif['radiant_top_gold_score'] = radiant_rank_gold_score.radiant_gold_score_rank_4
gold_stats_dif['radiant_second_gold_score'] = radiant_rank_gold_score.radiant_gold_score_rank_3
gold_stats_dif['dire_top_gold_score'] = dire_rank_gold_score.dire_gold_score_rank_4
gold_stats_dif['dire_second_gold_score'] = dire_rank_gold_score.dire_gold_score_rank_3

gold_stats_dif.reset_index(inplace=True)
gold_stats_dif.to_csv('processing_tables/gold_stat_dif.csv', index=None)
gold_stats_dif.head()
