==> 2 - Heroes.ipynb <==
import pandas as pd
import numpy as np

from sklearn.preprocessing import StandardScaler
from sklearn.grid_search import GridSearchCV
from sklearn.cross_validation import KFold
from sklearn.linear_model import LogisticRegression
from sklearn.cross_validation import cross_val_score
from sklearn.ensemble import RandomForestClassifier
from sklearn.linear_model import Lasso, Ridge

%pylab inline

def normalize_data(X):
    return pd.DataFrame(
        StandardScaler().fit_transform(X),
        index = X.index, 
        columns=X.columns
    )

def logEstimation(X, y):
    grid = {'C': np.power(10.0, np.arange(-5, 1))}
    kf=KFold(y.size, n_folds=5, shuffle=True, random_state=241)
    clf=LogisticRegression(random_state=241)
    gs = GridSearchCV(clf, grid, scoring='roc_auc', cv=kf)
    gs.fit(X, y)
    return gs

train_matches = pd.read_csv('data/train.csv')
test_matches = pd.read_csv('data/test.csv')

heroes = pd.read_csv('data/heroes.csv')
heroes.head()

players = {'player_{0}'.format(i) for i in range(10)}
print players

heroes.player_0.unique()

heroes_num = heroes.player_0.unique().max()
print 'Heroes num: ', heroes_num
X_pick = np.zeros((heroes.shape[0], heroes_num + 1))
print X_pick

for i, match_id in enumerate(heroes.mid.values):
    for p in xrange(5):
        X_pick[i, heroes.ix[match_id, 'player_{0}'.format(p)]] = 1
        X_pick[i, heroes.ix[match_id, 'player_{0}'.format(5 + p)]] = -1  

print X_pick

xpick_df = pd.DataFrame(X_pick, columns=['hero_{0}'.format(i) for i in range(111)])
xpick_df = xpick_df.astype(int)
xpick_df.index = heroes.index 

xpick_df['mid'] = xpick_df.index
xpick_df.head()

xpick_df.ix[0,:]

# add labels 

labels = pd.read_csv('processing_tables/heroes_labels.csv')
labels = labels.values.ravel()
print labels

from collections import Counter
dummy = pd.DataFrame(index=heroes.mid)
have_carry = []
have_support = []
to_many_carry = []
all_supps = []
for pick in heroes.ix[:, 1:6].values:
    current_game_labels = Counter([labels[hero] for hero in pick])
    have_carry.append(current_game_labels[3] > 0)
    have_support.append(current_game_labels[1] > 0)
    to_many_carry.append(current_game_labels[3] > 3)
    all_supps.append(current_game_labels[1] == 5)
dummy['Have_carry'] = have_carry
dummy['Have_support'] = have_support
dummy['To_many_carry'] = to_many_carry
dummy['All_supps'] = all_supps

for col in dummy.columns:
    dummy[col] = dummy[col].astype(int)
dummy.reset_index(inplace=True)
dummy.head()
dummy.to_csv('processing_tables/dummy_heroes.csv')

dummy.All_supps.head()

# Merge and save

xpick_df.to_csv('processing_tables/only_heroes.csv', index=None)
xpick_df_with_dummy = pd.merge(xpick_df, dummy, on='mid', how='left')
xpick_df_with_dummy.head()

xpick_df_with_dummy.to_csv('processing_tables/heroes_with_dummy.csv', index=None)

gold = pd.read_csv('processing_tables/goldScoreWithoutCarry.csv')
train = pd.merge(train_matches, gold, on='mid', how='left')
test = pd.merge(test_matches, gold, on='mid', how='left')
train = pd.merge(train, xpick_df, on='mid', how='left')
test = pd.merge(test, xpick_df, on='mid', how='left')

heroes.to_csv('processing_tables/only_heroes.csv', index=None)

train.to_csv('processing_tables/train_gold_heroes.csv', index=None)
test.to_csv('processing_tables/test_gold_heroes.csv', index=None)

# Modeling

X_train = train.drop('radiant_won', 1)
y_train = train.radiant_won
clf = RandomForestClassifier(n_estimators=500, max_features=2, max_depth=15, n_jobs=8, random_state=228)
np.mean(cross_val_score(clf, X_train, y_train, cv=5, scoring='roc_auc'))

gs = logEstimation(X_train, y_train)
print gs.grid_scores_
print gs.best_params_
print('best score: {}'.format(gs.best_score_))

clf = LogisticRegression(C=1)
clf.fit(X_train, y_train)
np.mean(cross_val_score(clf, X_train, y_train, cv=5, scoring='roc_auc'))

clf = Ridge(alpha=0.0001)
clf.fit(X_train, y_train)
np.mean(cross_val_score(clf, X_train, y_train, cv=5, scoring='roc_auc'))

clf = Lasso(alpha=0.0001, max_iter=6000)
clf.fit(X_train, y_train)
np.mean(cross_val_score(clf, X_train, y_train, cv=5, scoring='roc_auc'))

gold = pd.read_csv('processing_tables/goldScoreWithoutCarry.csv')
train = pd.merge(train_matches, gold, on='mid', how='left')
test = pd.merge(test_matches, gold, on='mid', how='left')
train = pd.merge(train, xpick_df, on='mid', how='left')
test = pd.merge(test, xpick_df, on='mid', how='left')
train = pd.merge(train, dummy, on='mid', how='left')
test = pd.merge(test, dummy, on='mid', how='left')
X_train = train.drop('radiant_won', 1)
X_train.drop('To_many_carry', 1, inplace=True)
y_train = train.radiant_won

clf = LogisticRegression(C=1)
np.mean(cross_val_score(clf, X_train, y_train, cv=5, scoring='roc_auc'))

clf = Ridge(alpha=0.0001)
clf.fit(X_train, y_train)
np.mean(cross_val_score(clf, X_train, y_train, cv=5, scoring='roc_auc'))

clf = Lasso(alpha=0.0001, max_iter=6000)
clf.fit(X_train, y_train)
np.mean(cross_val_score(clf, X_train, y_train, cv=5, scoring='roc_auc'))

# With gold score 

train = pd.read_csv('processing_tables/train_score_gold.csv')
test = pd.read_csv('processing_tables/test_score_gold.csv')
train = pd.merge(train, xpick_df, on='mid', how='left')
test = pd.merge(test, xpick_df, on='mid', how='left')
X_train = train.drop('radiant_won', 1)
y_train = train.radiant_won

train.to_csv('processing_tables/train_goldScore_heroes.csv', index=None)
test.to_csv('processing_tables/test_goldScore_heroes.csv', index=None)

clf = RandomForestClassifier(n_estimators=100, n_jobs=-1, random_state=228)
np.mean(cross_val_score(clf, X_train, y_train, cv=5, scoring='roc_auc'))

clf = RandomForestClassifier(n_estimators=500, max_features=2, max_depth=15, n_jobs=8, random_state=228)
np.mean(cross_val_score(clf, X_train, y_train, cv=5, scoring='roc_auc'))

gs = logEstimation(X_train, y_train)
print gs.grid_scores_
print gs.best_params_
print('best score: {}'.format(gs.best_score_))

clf = LogisticRegression(C=0.1)
np.mean(cross_val_score(clf, X_train, y_train, cv=5, scoring='roc_auc'))

clf = Ridge(alpha=0.0001)
np.mean(cross_val_score(clf, X_train, y_train, cv=5, scoring='roc_auc'))

clf = Lasso(alpha=0.0001, max_iter=4000)
np.mean(cross_val_score(clf, X_train, y_train, cv=5, scoring='roc_auc'))

clf = Lasso(alpha=0.0001, max_iter=6000)
np.mean(cross_val_score(clf, X_train, y_train, cv=5, scoring='roc_auc'))

# With score rank gold

train = pd.read_csv('processing_tables/train_scoreRankGold.csv')
test = pd.read_csv('processing_tables/test_scoreRankGold.csv')
train = pd.merge(train, xpick_df, on='mid', how='left')
test = pd.merge(test, xpick_df, on='mid', how='left')
X_train = train.drop('radiant_won', 1)
y_train = train.radiant_won

clf = RandomForestClassifier(n_estimators=100, n_jobs=-1, random_state=228)
np.mean(cross_val_score(clf, X_train, y_train, cv=5, scoring='roc_auc'))

gs = logEstimation(X_train, y_train)
print gs.grid_scores_
print gs.best_params_
print('best score: {}'.format(gs.best_score_))

clf = LogisticRegression(C=0.1)
np.mean(cross_val_score(clf, X_train, y_train, cv=5, scoring='roc_auc'))

clf = Ridge(alpha=0.0001)
np.mean(cross_val_score(clf, X_train, y_train, cv=5, scoring='roc_auc'))

clf = Lasso(alpha=0.0001, max_iter=6000)
np.mean(cross_val_score(clf, X_train, y_train, cv=5, scoring='roc_auc'))

train.to_csv('processing_tables/train_goldRankScore_heroes.csv', index=None)
test.to_csv('processing_tables/test_goldRankScore_heroes.csv', index=None)

# With norm gold and scores

train = pd.read_csv('processing_tables/train_ns_gold.csv')
test = pd.read_csv('processing_tables/test_ns_gold.csv')
train = pd.merge(train, xpick_df, on='mid', how='left')
test = pd.merge(test, xpick_df, on='mid', how='left')
X_train = train.drop('radiant_won', 1)
y_train = train.radiant_won

gs = logEstimation(X_train, y_train)
print gs.grid_scores_
print gs.best_params_
print('best score: {}'.format(gs.best_score_))

clf = LogisticRegression(C=1)
clf.fit(X_train, y_train)
np.mean(cross_val_score(clf, X_train, y_train, cv=5, scoring='roc_auc'))

clf = Ridge(alpha=0.0001)
clf.fit(X_train, y_train)
np.mean(cross_val_score(clf, X_train, y_train, cv=5, scoring='roc_auc'))

clf = Lasso(alpha=0.0001, max_iter=4000)
clf.fit(X_train, y_train)
np.mean(cross_val_score(clf, X_train, y_train, cv=5, scoring='roc_auc'))

X_train.drop(['diff_gold', 'ratio_gold'], 1, inplace=True)

gs = logEstimation(X_train, y_train)
print gs.grid_scores_
print gs.best_params_
print('best score: {}'.format(gs.best_score_))

clf = LogisticRegression(C=0.1)
clf.fit(X_train, y_train)
np.mean(cross_val_score(clf, X_train, y_train, cv=5, scoring='roc_auc'))

clf = Ridge(alpha=0.0001)
clf.fit(X_train, y_train)
np.mean(cross_val_score(clf, X_train, y_train, cv=5, scoring='roc_auc'))

clf = Lasso(alpha=0.0001, max_iter=4000)
clf.fit(X_train, y_train)
np.mean(cross_val_score(clf, X_train, y_train, cv=5, scoring='roc_auc'))
