==> 3 - Items.ipynb <==
import pandas as pd
import numpy as np

from sklearn.preprocessing import StandardScaler
from sklearn.grid_search import GridSearchCV
from sklearn.cross_validation import KFold
from sklearn.linear_model import LogisticRegression
from sklearn.cross_validation import cross_val_score
from sklearn.ensemble import RandomForestClassifier
from sklearn.linear_model import Lasso, Ridge

%pylab inline

def normalize_data(X):
    return pd.DataFrame(
        StandardScaler().fit_transform(X),
        index = X.index, 
        columns=X.columns
    )

def logEstimation(X, y):
    grid = {'C': np.power(10.0, np.arange(-5, 1))}
    kf=KFold(y.size, n_folds=5, shuffle=True, random_state=241)
    clf=LogisticRegression(random_state=241)
    gs = GridSearchCV(clf, grid, scoring='roc_auc', cv=kf)
    gs.fit(X, y)
    return gs

items = pd.read_csv('data/items.csv')
items.fillna(0, inplace=True)
items.head()

radiant_items = items.drop('player', 1).loc[items.player < 5].groupby('mid').sum()
norm_radiant_items = pd.DataFrame(index=radiant_items.index)
for col in radiant_items.columns:
    norm_radiant_items['rad_' + col] = radiant_items[col].astype(int32)
norm_radiant_items.reset_index(level=0, inplace=True)
norm_radiant_items.head()

### Возможно стоило бы отмасштабировать, но не буду

dire_items = items.drop('player', 1).loc[items.player > 4].groupby('mid').sum()
norm_dire_items = pd.DataFrame(index=dire_items.index)
for col in dire_items.columns:
    norm_dire_items['dire_' + col] = dire_items[col].astype(int32)
norm_dire_items.reset_index(level=0, inplace=True)
norm_dire_items.head()

new_items = pd.merge(norm_radiant_items, norm_dire_items, on='mid', how='left')
new_items.head()

new_items.to_csv('processing_tables/only_items.csv', index=None)

# Items dif

items = pd.read_csv('data/items.csv')
items.fillna(0, inplace=True)
radiant_items = items.drop('player', 1).loc[items.player < 5].groupby('mid').sum()
dire_items = items.drop('player', 1).loc[items.player > 4].groupby('mid').sum()

items_dif = pd.DataFrame(index=dire_items.index)
for col in dire_items.columns:
    items_dif[col + '_dif'] = (radiant_items[col] - dire_items[col]).astype(int32)
items_dif.reset_index(level=0, inplace=True)
items_dif.to_csv('processing_tables/items_dif.csv', index=None)
items_dif.head()

# Merge and save

train = pd.read_csv('processing_tables/train_gold_heroes.csv')
test = pd.read_csv('processing_tables/test_gold_heroes.csv')

train = pd.merge(train, new_items, on='mid', how='left')
test = pd.merge(test, new_items, on='mid', how='left')

new_items.to_csv('processing_tables/processed_items.csv', index=None)
train.to_csv('processing_tables/train_gold_heroes_items.csv', index=None)
test.to_csv('processing_tables/test_gold_heroes_items.csv', index=None)
