==> 4 - Experiance.ipynb <==
import pandas as pd
import numpy as np

from sklearn.preprocessing import StandardScaler
from sklearn.grid_search import GridSearchCV
from sklearn.cross_validation import KFold
from sklearn.linear_model import LogisticRegression
from sklearn.cross_validation import cross_val_score
from sklearn.ensemble import RandomForestClassifier
from sklearn.linear_model import Lasso, Ridge

%pylab inline

def normalize_data(X):
    return pd.DataFrame(
        StandardScaler().fit_transform(X),
        index = X.index, 
        columns=X.columns
    )

def logEstimation(X, y):
    grid = {'C': np.power(10.0, np.arange(-5, 1))}
    kf=KFold(y.size, n_folds=5, shuffle=True, random_state=241)
    clf=LogisticRegression(random_state=241)
    gs = GridSearchCV(clf, grid, scoring='roc_auc', cv=kf)
    gs.fit(X, y)
    return gs

# Norm xp

xp = pd.read_csv('data/xp.csv', index_col='mid')
xp = xp.loc[xp.times == 600].drop('times', 1)

radiant_xp = xp[['player_0', 'player_1', 'player_2', 'player_3', 'player_4']].sum(axis=1)
dire_xp = xp[['player_5', 'player_6', 'player_7', 'player_8', 'player_9']].sum(axis=1)
radiant_carry_xp = xp[['player_0', 'player_1', 'player_2', 'player_3', 'player_4']].max(axis=1)
dire_carry_xp = xp[['player_5', 'player_6', 'player_7', 'player_8', 'player_9']].max(axis=1)
norm_xp = pd.DataFrame(index=xp.index)
for col in xp.columns:
    norm_xp['exp_' + col] = xp[col]
    
norm_xp['radiant_xp'] = radiant_xp
norm_xp['dire_xp'] = dire_xp
norm_xp['radiant_carry_xp'] = radiant_carry_xp
norm_xp['dire_carry_xp'] = dire_carry_xp

norm_xp = normalize_data(norm_xp)
norm_xp.reset_index(level=0, inplace=True)
norm_xp.head()

norm_xp.to_csv('processing_tables/only_exp.csv', index=None)

xp = pd.read_csv('data/xp.csv', index_col='mid')
xp = xp[xp.times == 600]
xp.drop('times', 1, inplace=True)
radiant_rank_xp = pd.DataFrame(data = np.sort(xp.values[:,:5]),
                                 index=xp.index,
                                 columns = ['radiant_xp_rank_{0}'.format(i) for i in range(5)]) 
radiant_rank_xp.reset_index(inplace=True)

dire_rank_xp = pd.DataFrame(data = np.sort(xp.values[:,5:10]),
                                 index=xp.index,
                                 columns = ['dire_xp_rank_{0}'.format(i) for i in range(5)]) 

dire_rank_xp.reset_index(inplace=True)
dire_rank_xp.head()

xp_stat = pd.DataFrame(index=xp.index)
xp_stat['rad_best_xp'] = radiant_rank_xp.radiant_xp_rank_4
xp_stat['rad_second_xp'] = radiant_rank_xp.radiant_xp_rank_3

xp_stat['dire_best_xp'] = dire_rank_xp.dire_xp_rank_4
xp_stat['dire_second_xp'] = dire_rank_xp.dire_xp_rank_3

xp_stat['rad_sum_xp'] = xp[['player_0', 'player_1', 'player_2', 'player_3', 'player_4']].sum(axis=1)
xp_stat['dire_sum_xp'] = xp[['player_5', 'player_6', 'player_7', 'player_8', 'player_9']].sum(axis=1)

# xp_stat = normalize_data(xp_stat)
xp_stat.reset_index(inplace=True)
# xp_stat.to_csv('processing_tables/xp_stat.csv', index=None)
xp_stat.head()


xp_cropped_stats = pd.DataFrame(index=xp.index)
xp_cropped_stats['xp_best_dif'] = xp_stat.rad_best_xp - xp_stat.dire_best_xp
xp_cropped_stats['xp_second_dif'] = xp_stat.rad_second_xp - xp_stat.dire_second_xp
xp_cropped_stats['xp_sum_dif'] = xp_stat.rad_sum_xp - xp_stat.dire_sum_xp
xp_cropped_stats = normalize_data(xp_cropped_stats)
xp_cropped_stats.reset_index(inplace=True)
xp_cropped_stats.to_csv('processing_tables/xp_cropped_stats.csv', index=None)
xp_cropped_stats.head()

# Exp Score

new_xp = pd.read_csv('data/xp.csv', index_col='mid')
new_xp = new_xp[new_xp.times == 600]
new_xp.drop('times', 1, inplace=True)
new_xp.head()

heroes = pd.read_csv('data/heroes.csv', index_col='mid')
heroes_num = heroes.max()[0] + 1
print heroes_num

mean_xp = np.zeros(111)
for hero_num in range(111):
    hero_values = []
    for player_num in range(10):
        colname = 'player_{}'.format(player_num)
        hero_index = heroes.loc[heroes[colname] == hero_num].index
        hero_values.extend(new_xp.loc[hero_index][colname].values)
    mean_xp[hero_num] = int(np.mean(hero_values))
mean_xp[0] = 100000
print mean_xp

Hero = 0 - соответствует вышедшему из игры человеку, поэтому сделаем соответствующее значение побольше

mean_predict_xp = heroes.apply(lambda x: mean_xp[x])
mean_predict_xp.head()

temp_xp_score = new_xp / mean_predict_xp
xp_score = pd.DataFrame(index=temp_xp_score.index)
for col in temp_xp_score.columns:
    xp_score['xp_score_' + col] = temp_xp_score[col]

# xp_score['radiant_norm_xp'] = norm_xp.radiant_xp
# xp_score['dire_norm_xp'] = norm_xp.dire_xp
# xp_score['radiant_carry_norm_xp'] = norm_xp.radiant_carry_xp
# xp_score['dire_carry_norm_xp'] = norm_xp.dire_carry_xp

xp_score['radiant_best_xp_score'] = xp_score.iloc[:,1:6].max(1)
xp_score['dire_best_xp_score'] = xp_score.iloc[:,6:11].max(1)
xp_score['radiant_sum_xp_score'] = StandardScaler().fit_transform(xp_score.iloc[:,1:6].sum(1).reshape(-1, 1))
xp_score['dire_sum_xp_score'] = StandardScaler().fit_transform(xp_score.iloc[:,6:11].sum(1).reshape(-1, 1))

xp_score.reset_index(level=0, inplace=True)
xp_score.head()

xp_score.to_csv('processing_tables/xp_score.csv', index=None)

pd.merge(xp_score, norm_xp, on='mid').head()

# rank xp dif

xp = pd.read_csv('data/xp.csv', index_col='mid')
xp = xp[xp.times == 600]
xp.drop('times', 1, inplace=True)
xp.head()

rank_xp_dif = pd.DataFrame(data = np.sort(xp.values[:,:5]) - np.sort(xp.values[:,5:10]),
                                 index=xp.index,
                                 columns = ['xp_dif_{0}'.format(i) for i in range(5)]) 
rank_xp_dif = normalize_data(rank_xp_dif)
rank_xp_dif.reset_index(inplace=True)

rank_xp_dif.head()

rank_xp_dif.to_csv('processing_tables/xpDif.csv', index=None)
