==> 5 - CreepStats.ipynb <==
import pandas as pd
import numpy as np

from sklearn.preprocessing import StandardScaler
from sklearn.grid_search import GridSearchCV
from sklearn.cross_validation import KFold
from sklearn.linear_model import LogisticRegression
from sklearn.cross_validation import cross_val_score
from sklearn.ensemble import RandomForestClassifier
from sklearn.linear_model import Lasso, Ridge

%pylab inline

def normalize_data(X):
    return pd.DataFrame(
        StandardScaler().fit_transform(X),
        index = X.index, 
        columns=X.columns
    )

def logEstimation(X, y):
#     grid = {'C': np.power(10.0, np.arange(-5, 1))}
    grid = {'C': [0.05, 0.01, 0.05, 0.1, 0.5, 0.75, 1]}
    kf=KFold(y.size, n_folds=5, shuffle=True, random_state=241)
    clf=LogisticRegression(random_state=241)
    gs = GridSearchCV(clf, grid, scoring='roc_auc', cv=kf)
    gs.fit(X, y)
    return gs

lh = pd.read_csv('data\lh.csv', index_col='mid')
lh = lh.loc[lh.times == 600].drop('times', 1)
lh.head()

radiant_lh = lh.player_0 + lh.player_1 + lh.player_2 + lh.player_3 + lh.player_4
dire_lh = lh.player_5 + lh.player_6 + lh.player_7 + lh.player_8 + lh.player_9
radiant_carry_cs = lh[['player_0', 'player_1', 'player_2', 'player_3', 'player_4']].max(1)
dire_carry_cs = lh[['player_5', 'player_6', 'player_7', 'player_8', 'player_9']].max(1)
creeps = pd.DataFrame(index=lh.index)
for col in lh.columns:
    creeps['creeps_' + col] = lh[col]
creeps['radiant_creeps'] = radiant_lh
creeps['dire_creeps'] = dire_lh
creeps['radiant_carry_creeps'] = radiant_carry_cs
creeps['dire_carry_creeps'] = dire_carry_cs
creeps['creeps_diff'] = radiant_lh - dire_lh
creeps.head()

creeps = normalize_data(creeps)
creeps['creeps_rel'] = radiant_lh / dire_lh
creeps.reset_index(level=0, inplace=True)

creeps.head()

creeps.to_csv('processing_tables/creeps.csv', index=None)

creeps = pd.read_csv('data\lh.csv', index_col='mid')
creeps = creeps.loc[creeps.times == 600].drop('times', 1)
creeps.reset_index(inplace=True)
radiant_rank_creep = pd.DataFrame(data = np.sort(creeps.values[:,1:6]),
                                       index=creeps.mid,
                                       columns = ['radiant_creep_rank_{0}'.format(i) for i in range(5)]) 
radiant_rank_creep = normalize_data(radiant_rank_creep)
dire_rank_creep = pd.DataFrame(data = np.sort(creeps.values[:,6:11]),
                                       index=creeps.mid,
                                       columns = ['dire_creep_rank_{0}'.format(i) for i in range(5)]) 
dire_rank_creep = normalize_data(dire_rank_creep)

cropped_creeps = pd.DataFrame(index = creeps.mid)
cropped_creeps['rad_best_creeps'] = radiant_rank_creep.radiant_creep_rank_4
cropped_creeps['rad_second_creeps'] = radiant_rank_creep.radiant_creep_rank_3
cropped_creeps['dire_best_creeps'] = dire_rank_creep.dire_creep_rank_4
cropped_creeps['dire_second_creeps'] = dire_rank_creep.dire_creep_rank_3
cropped_creeps['rad_creeps_sum'] = radiant_lh
cropped_creeps['dire_creeps_sum'] = dire_lh
cropped_creeps = normalize_data(cropped_creeps)
cropped_creeps.reset_index(inplace=True)

cropped_creeps.head()

cropped_creeps.to_csv('processing_tables/cropped_creeps2.csv', index=None)

creeps = pd.read_csv('data\lh.csv', index_col='mid')
creeps = creeps.loc[creeps.times == 600].drop('times', 1)
creeps.reset_index(inplace=True)
radiant_rank_creep = pd.DataFrame(data = np.sort(creeps.values[:,1:6]),
                                       index=creeps.mid,
                                       columns = ['radiant_creep_rank_{0}'.format(i) for i in range(5)]) 
# radiant_rank_creep = normalize_data(radiant_rank_creep)
dire_rank_creep = pd.DataFrame(data = np.sort(creeps.values[:,6:11]),
                                       index=creeps.mid,
                                       columns = ['dire_creep_rank_{0}'.format(i) for i in range(5)]) 
# dire_rank_creep = normalize_data(dire_rank_creep)

cropped_creeps = pd.DataFrame(index = creeps.mid)
cropped_creeps['rad_best_creeps'] = radiant_rank_creep.radiant_creep_rank_4
cropped_creeps['rad_second_creeps'] = radiant_rank_creep.radiant_creep_rank_3
cropped_creeps['dire_best_creeps'] = dire_rank_creep.dire_creep_rank_4
cropped_creeps['dire_second_creeps'] = dire_rank_creep.dire_creep_rank_3
cropped_creeps['rad_creeps_sum'] = creeps.ix[:, 0:5].sum(1)
cropped_creeps['dire_creeps_sum'] = creeps.ix[:, 5:10].sum(1)
# cropped_creeps = normalize_data(cropped_creeps)
cropped_creeps.reset_index(inplace=True)

creeps_dif = pd.DataFrame(index = creeps.mid)
creeps_dif['best_creeps_dif'] = cropped_creeps.rad_best_creeps - cropped_creeps.dire_best_creeps
creeps_dif['second_creeps_dif'] = cropped_creeps.rad_second_creeps - cropped_creeps.dire_second_creeps
creeps_dif['sum_creeps_dif'] = cropped_creeps.rad_creeps_sum - cropped_creeps.dire_creeps_sum
# creeps_dif = normalize_data(creeps_dif)
creeps_dif.reset_index(inplace=True)
creeps_dif.head()

creeps_dif.to_csv('processing_tables/cropped_creeps_dif.csv', index=None)

# Merge & Save

train = pd.read_csv('processing_tables/train_gold_heroes_items_exp.csv')
test = pd.read_csv('processing_tables/test_gold_heroes_items_exp.csv')

train = pd.merge(train, creeps, on='mid', how='left')
test = pd.merge(test, creeps, on='mid', how='left')

creeps.to_csv('processing_tables/processed_creeps.csv', index=None)
train.to_csv('processing_tables/train_gold_heroes_items_exp_creeps.csv', index=None)
test.to_csv('processing_tables/test_gold_heroes_items_exp_creeps.csv', index=None)

# Modeling

X_train = train.drop('radiant_won', 1)
y_train = train.radiant_won

gs = logEstimation(X_train, y_train)
print gs.grid_scores_
print gs.best_params_
print('best score: {}'.format(gs.best_score_))

clf = LogisticRegression(C=0.5)
np.mean(cross_val_score(clf, X_train, y_train, cv=5, scoring='roc_auc'))

clf = LogisticRegression(C=0.1)
np.mean(cross_val_score(clf, X_train, y_train, cv=5, scoring='roc_auc'))

clf = Ridge(alpha=0.0001)
np.mean(cross_val_score(clf, X_train, y_train, cv=5, scoring='roc_auc'))

X_train = train.drop(['radiant_won', 'radiant_xp', 'dire_xp'], 1)
clf = Lasso(alpha=0.0001)
clf.fit(X_train, y_train)
np.mean(cross_val_score(clf, X_train, y_train, cv=5, scoring='roc_auc'))

X_train = train.drop(['radiant_won', 'radiant_xp', 'dire_xp'], 1)
clf = Lasso(alpha=0.0001)
clf.fit(X_train, y_train)

X_test = test.drop(['radiant_xp', 'dire_xp'], 1)
test_matches = pd.read_csv('data/test.csv')
test_matches['radiant_won'] = clf.predict(X_test)

test_matches.head()

test_matches.to_csv('submissions/with_creeps.csv', index=None)

# With gold score

train = pd.read_csv('processing_tables/train_goldScore_heroes_items_exp.csv')
test = pd.read_csv('processing_tables/test_goldScore_heroes_items_exp.csv')

train = pd.merge(train, creeps, on='mid', how='left')
test = pd.merge(test, creeps, on='mid', how='left')

X_train = train.drop('radiant_won', 1)
y_train = train.radiant_won

gs = logEstimation(X_train, y_train)
print gs.grid_scores_
print gs.best_params_
print('best score: {}'.format(gs.best_score_))

clf = LogisticRegression(C=0.75)
np.mean(cross_val_score(clf, X_train, y_train, cv=5, scoring='roc_auc'))

clf = Ridge(alpha=0.0001)
np.mean(cross_val_score(clf, X_train, y_train, cv=5, scoring='roc_auc'))

clf = Lasso(alpha=0.0001, max_iter=6000)
np.mean(cross_val_score(clf, X_train, y_train, cv=5, scoring='roc_auc'))

creeps.to_csv('processing_tables/processed_creeps.csv', index=None)
train.to_csv('processing_tables/train_goldScore_heroes_items_exp_creeps.csv', index=None)
test.to_csv('processing_tables/test_goldScore_heroes_items_exp_creeps.csv', index=None)

# Creep_score

new_lh = pd.read_csv('data\lh.csv', index_col='mid')
new_lh = new_lh.loc[new_lh.times == 600].drop('times', 1)
new_lh.head()

heroes = pd.read_csv('data/heroes.csv', index_col='mid')
heroes_num = heroes.max()[0] + 1
print heroes_num

mean_creep_stat = np.zeros(111)
for hero_num in range(111):
    hero_values = []
    for player_num in range(10):
        colname = 'player_{}'.format(player_num)
        hero_index = heroes.loc[heroes[colname] == hero_num].index
        hero_values.extend(new_lh.loc[hero_index][colname].values)
    mean_creep_stat[hero_num] = int(np.mean(hero_values))
mean_creep_stat[0] = 100000
mean_creep_stat = np.array(map(int, mean_creep_stat))
print mean_creep_stat
